+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Released on Sunday, July 26 2026.
 - Added missing scroll declarations (e.g., `scroll-behavior`, `overscroll-behavior`)
 - Added missing fonts declarations (e.g., `font-display`, `font-palette`, `font-synthesis`)
 - Added missing color declarations (e.g., `color-scheme`, `accent-color`)
+- Added missing compositing declarations (e.g., `backdrop-filter`, `isolation`)
 - Added more CSSOM possibilities and helpers (#6)
 - Added parts of recent color spec update such as `rgb` with spaces (#131)
 - Added now Color L4 parsing with `hsl`, `hwb`, `lab`, `lch`, `oklab`, and `oklch`